+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The following changes have been completed, and will be included in the next tagg
 * Added `PackageName` type.
 * Added `isPackageNameData()` data guard.
 * Added `mustBePackageNameData()` data guarantee.
+* Added `packageNameFrom()` smart constructor.
 
 ### Fixes